requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
statsmodels==0.14.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Excel/Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# API Development
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.1.1

# Workflow & Automation
schedule==1.2.0
python-dotenv==1.0.0

# Configuration Management
pyyaml==6.0
jsonschema==4.18.4

# Logging & Monitoring
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Notebook Support
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.0

# Utilities
tqdm==4.65.0
python-dateutil==2.8.2
pytz==2023.3

# Version Control for Models
dvc==3.15.0

# Optional: Dashboard
streamlit==1.25.0
